move events package to ecs package
because that is where it belongs
